separated test_app into compress and decompress
