+++ examples/compress.rs
@@ -3,7 +3,7 @@ extern crate gfwx;
 extern crate image;
 extern crate time;
 
-use std::{error::Error, fs, i64, io, io::prelude::*, path::Path};
+use std::{error::Error, fs, i64, io::prelude::*, path::Path};
 
 use image::DynamicImage::*;
 use time::PreciseTime;
@@ -11,12 +11,9 @@ use time::PreciseTime;
 fn main() -> Result<(), Box<dyn Error>> {
     let matches = get_matches();
     let input_file = matches.value_of("INPUT").unwrap();
-    let output_gfwx = matches.value_of("OUTPUT.GFWX").unwrap();
-    let output_decompressed = matches.value_of("OUTPUT").unwrap();
+    let output_gfwx = matches.value_of("OUTPUT").unwrap();
     let quality = matches.value_of("quality").unwrap().parse().unwrap();
-    let chroma_scale = matches.value_of("chroma-scale").unwrap().parse().unwrap();
     let block_size = matches.value_of("block-size").unwrap().parse().unwrap();
-    let downsampling = matches.value_of("downsampling").unwrap().parse().unwrap();
     let filter = match matches.value_of("filter").unwrap() {
         "linear" => gfwx::Filter::Linear,
         "cubic" => gfwx::Filter::Cubic,
@@ -28,6 +25,7 @@ fn main() -> Result<(), Box<dyn Error>> {
         "contextual" => gfwx::Encoder::Contextual,
         _ => panic!("clap betrayed us again"),
     };
+
     let file_path = Path::new(&input_file);
     let image = image::open(&file_path)?;
     let (width, height, image, channels, intent) =
@@ -42,7 +40,7 @@ fn main() -> Result<(), Box<dyn Error>> {
         bit_depth: 8,
         is_signed: false,
         quality,
-        chroma_scale,
+        chroma_scale: 8,
         block_size,
         filter,
         quantization: gfwx::Quantization::Scalar,
@@ -52,10 +50,11 @@ fn main() -> Result<(), Box<dyn Error>> {
     };
 
     let mut compressed = vec![0; 2 * image.len()];
+
     let compress_start = PreciseTime::now();
     let gfwx_size = gfwx::compress_simple(&image, &header, &gfwx::ColorTransformProgram::new(), &mut compressed)?;
     let compress_end = PreciseTime::now();
-    compressed.truncate(gfwx_size);
+
     println!(
         "Compression took {} microseconds",
         compress_start
@@ -63,32 +62,9 @@ fn main() -> Result<(), Box<dyn Error>> {
             .num_microseconds()
             .unwrap_or(i64::MAX)
     );
-    {
-        let mut f = fs::File::create(output_gfwx)?;
-        f.write_all(&compressed)?;
-    }
 
-    let mut slice = compressed.as_slice();
-    let header = gfwx::Header::decode(&mut slice).unwrap();
-    let mut decompressed = vec![0; header.get_decompress_buffer_size(downsampling).unwrap()];
-    let decompress_start = PreciseTime::now();
-    gfwx::decompress_simple(&slice, &header, downsampling, &mut decompressed)?;
-    let decompress_end = PreciseTime::now();
-    println!(
-        "Decompression took {} microseconds",
-        decompress_start
-            .to(decompress_end)
-            .num_microseconds()
-            .unwrap_or(i64::MAX)
-    );
-
-    save_image(
-        decompressed,
-        intent,
-        header.get_downsampled_width(downsampling) as u32,
-        header.get_downsampled_height(downsampling) as u32,
-        &output_decompressed,
-    )?;
+    let mut f = fs::File::create(output_gfwx)?;
+    f.write_all(&compressed[0..gfwx_size])?;
 
     Ok(())
 }
@@ -104,16 +80,10 @@ fn get_matches() -> clap::ArgMatches<'static> {
                 .index(1),
         )
         .arg(
-            clap::Arg::with_name("OUTPUT.GFWX")
+            clap::Arg::with_name("OUTPUT")
                 .help("Sets the output file to write compressed gfwx")
                 .required(true)
                 .index(2),
-        )
-        .arg(
-            clap::Arg::with_name("OUTPUT")
-                .help("Sets the output file to write decompressed image")
-                .required(true)
-                .index(3),
         )
         .arg(
             clap::Arg::with_name("quality")
@@ -129,29 +99,6 @@ fn get_matches() -> clap::ArgMatches<'static> {
                     } else {
                         Err("Quality must be in range (1..=1024)".to_string())
                     }
-                }),
-        )
-        .arg(
-            clap::Arg::with_name("chroma-scale")
-                .help("Sets the chroma scale for compression")
-                .short("c")
-                .long("chroma-scale")
-                .takes_value(true)
-                .default_value("8")
-                .validator(|v| {
-                    v.parse::<u8>().map_err(|e| e.to_string())?;
-                    Ok(())
-                }),
-        )
-        .arg(
-            clap::Arg::with_name("downsampling")
-                .help("Sets the downsampling scale for decompression")
-                .short("d") .long("downsampling")
-                .takes_value(true)
-                .default_value("0")
-                .validator(|v| {
-                    v.parse::<u8>().map_err(|e| e.to_string())?;
-                    Ok(())
                 }),
         )
         .arg(
@@ -184,7 +131,7 @@ fn get_matches() -> clap::ArgMatches<'static> {
                 .short("e")
                 .long("encoder")
                 .takes_value(true)
-                .default_value("contextual")
+                .default_value("turbo")
                 .possible_values(&["turbo", "fast", "contextual"]),
         )
         .arg(
@@ -229,36 +176,5 @@ fn into_raw_image(
             ImageBgra8(i) => (i.width(), i.height(), i.into_raw(), 4, gfwx::Intent::BGRA),
             _ => panic!("unsupported image color space"),
         },
-    }
-}
-
-fn save_image(
-    decompressed: Vec<u8>,
-    intent: gfwx::Intent,
-    width: u32,
-    height: u32,
-    path: &AsRef<Path>,
-) -> io::Result<()> {
-    let decompressed_image = match intent {
-        gfwx::Intent::RGB => {
-            ImageRgb8(image::RgbImage::from_raw(width, height, decompressed).unwrap())
-        }
-        gfwx::Intent::RGBA => {
-            ImageRgba8(image::RgbaImage::from_raw(width, height, decompressed).unwrap())
-        }
-        gfwx::Intent::BGR => ImageBgr8(
-            image::ImageBuffer::<image::Bgr<u8>, Vec<u8>>::from_raw(width, height, decompressed)
-                .unwrap(),
-        ),
-        gfwx::Intent::BGRA => ImageBgra8(
-            image::ImageBuffer::<image::Bgra<u8>, Vec<u8>>::from_raw(width, height, decompressed)
-                .unwrap(),
-        ),
-        _ => panic!("Unsupported image intent"),
-    };
-    match intent {
-        gfwx::Intent::BGR => decompressed_image.to_rgb().save(&path),
-        gfwx::Intent::BGRA => decompressed_image.to_rgba().save(&path),
-        _ => decompressed_image.save(&path),
     }
 }
